Move specs across for DeprecatedReferencesLoader
